requirements.txt
fastapi
uvicorn
django
django-tailwind
pymysql
sqlalchemy
python-dotenv
apscheduler
requests
httpx
